+++ crypto/keyring/autocli.go
@@ -2,9 +2,9 @@ package keyring
 
 import (
 	signingv1beta1 "cosmossdk.io/api/cosmos/tx/signing/v1beta1"
-	authsigning "cosmossdk.io/x/auth/signing"
 
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 )
 
 // autoCLIKeyring represents the keyring interface used by the AutoCLI.