+++ runtime/gas.go
@@ -36,6 +36,10 @@ func (cgm CoreGasmeter) Consume(amount gas.Gas, descriptor string) error {
 	return nil
 }
 
+func (cgm CoreGasmeter) Consumed() gas.Gas {
+	return cgm.gm.GasConsumed()
+}
+
 func (cgm CoreGasmeter) Refund(amount gas.Gas, descriptor string) error {
 	cgm.gm.RefundGas(amount, descriptor)
 	return nil