+++ schema/testing/field_test.go
@@ -5,12 +5,14 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"pgregory.net/rapid"
+
+	"cosmossdk.io/schema"
 )
 
 func TestField(t *testing.T) {
 	rapid.Check(t, func(t *rapid.T) {
-		field := FieldGen.Draw(t, "field")
-		require.NoError(t, field.Validate())
+		field := FieldGen(testEnumSchema).Draw(t, "field")
+		require.NoError(t, field.Validate(testEnumSchema))
 	})
 }
 
@@ -19,8 +21,13 @@ func TestFieldValue(t *testing.T) {
 }
 
 var checkFieldValue = func(t *rapid.T) {
-	field := FieldGen.Draw(t, "field")
-	require.NoError(t, field.Validate())
-	fieldValue := FieldValueGen(field).Draw(t, "fieldValue")
-	require.NoError(t, field.ValidateValue(fieldValue))
+	field := FieldGen(testEnumSchema).Draw(t, "field")
+	require.NoError(t, field.Validate(testEnumSchema))
+	fieldValue := FieldValueGen(field, testEnumSchema).Draw(t, "fieldValue")
+	require.NoError(t, field.ValidateValue(fieldValue, testEnumSchema))
 }
+
+var testEnumSchema = schema.MustCompileModuleSchema(schema.EnumType{
+	Name:   "test_enum",
+	Values: []schema.EnumValueDefinition{{Name: "a", Value: 1}, {Name: "b", Value: 2}},
+})