+++ x/auth/ante/ante.go
@@ -4,12 +4,13 @@ import (
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/gas"
 	errorsmod "cosmossdk.io/errors"
-	"cosmossdk.io/x/auth/types"
 	txsigning "cosmossdk.io/x/tx/signing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
+	"github.com/cosmos/cosmos-sdk/x/auth/ante/unorderedtx"
+	"github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
 // HandlerOptions are the options required for constructing a default SDK AnteHandler.
@@ -23,6 +24,7 @@ type HandlerOptions struct {
 	SignModeHandler          *txsigning.HandlerMap
 	SigGasConsumer           func(meter gas.Meter, sig signing.SignatureV2, params types.Params) error
 	TxFeeChecker             TxFeeChecker
+	UnorderedTxManager       *unorderedtx.Manager
 }
 
 // NewAnteHandler returns an AnteHandler that checks and increments sequence
@@ -51,6 +53,10 @@ func NewAnteHandler(options HandlerOptions) (sdk.AnteHandler, error) {
 		NewDeductFeeDecorator(options.AccountKeeper, options.BankKeeper, options.FeegrantKeeper, options.TxFeeChecker),
 		NewValidateSigCountDecorator(options.AccountKeeper),
 		NewSigVerificationDecorator(options.AccountKeeper, options.SignModeHandler, options.SigGasConsumer, options.AccountAbstractionKeeper),
+	}
+
+	if options.UnorderedTxManager != nil {
+		anteDecorators = append(anteDecorators, NewUnorderedTxDecorator(unorderedtx.DefaultMaxTimeoutDuration, options.UnorderedTxManager, options.Environment, DefaultSha256Cost))
 	}
 
 	return sdk.ChainAnteDecorators(anteDecorators...), nil