+++ x/staking/keeper/val_state_change.go
@@ -266,14 +266,22 @@ func (k Keeper) ApplyAndReturnValidatorSetUpdates(ctx context.Context) ([]appmod
 			return nil, err
 		}
 
-		oldPk, ok := history.OldConsPubkey.GetCachedValue().(cryptotypes.PubKey)
+		oldPkCached := history.OldConsPubkey.GetCachedValue()
+		if oldPkCached == nil {
+			return nil, errorsmod.Wrap(sdkerrors.ErrInvalidType, "OldConsPubkey cached value is nil")
+		}
+		oldPk, ok := oldPkCached.(cryptotypes.PubKey)
 		if !ok {
-			return nil, errorsmod.Wrapf(sdkerrors.ErrInvalidType, "Expecting cryptotypes.PubKey, got %T", oldPk)
+			return nil, errorsmod.Wrapf(sdkerrors.ErrInvalidType, "Expecting cryptotypes.PubKey, got %T", oldPkCached)
 		}
 
-		newPk, ok := history.NewConsPubkey.GetCachedValue().(cryptotypes.PubKey)
+		newPkCached := history.NewConsPubkey.GetCachedValue()
+		if newPkCached == nil {
+			return nil, errorsmod.Wrap(sdkerrors.ErrInvalidType, "NewConsPubkey cached value is nil")
+		}
+		newPk, ok := newPkCached.(cryptotypes.PubKey)
 		if !ok {
-			return nil, errorsmod.Wrapf(sdkerrors.ErrInvalidType, "Expecting cryptotypes.PubKey, got %T", newPk)
+			return nil, errorsmod.Wrapf(sdkerrors.ErrInvalidType, "Expecting cryptotypes.PubKey, got %T", newPkCached)
 		}
 
 		// a validator cannot rotate keys if it's not bonded or if it's jailed